indicate private and public routes

+++ src/app/modules/subject/subject.service.ts
@@ -1,3 +1,5 @@
+import { AppError } from "../../helpers/appError";
+import { IRequestUser } from "../../interface/requestUser.interface";
 import { prisma } from "../../lib/prisma";
 import { ISubjectPayload } from "./subject.interface";
 
@@ -11,29 +13,72 @@ const getAllSubject = async () => {
 };
 
 
-const createSubject = async (payload: ISubjectPayload) => {
+const createSubject = async (user: IRequestUser, payload: ISubjectPayload) => {
+
+  const tutor = await prisma.tutor.findUnique({
+    where: { userId: user.userId }
+  });
+
+  if (!tutor) {
+    throw new AppError("Tutor not found", 404);
+  }
+
   return await prisma.tutorSubject.create({
-    data: payload,
+    data: {
+      tutorId: tutor.id,
+      categoryId: payload.categoryId,
+    },
     include: {
       tutor: true,
       category: true
-    },
+    }
   });
 };
 
-const updateSubject = async (id: string, data: ISubjectPayload) => {
+const updateSubject = async (user: IRequestUser, id: string, payload: ISubjectPayload) => {
+  const tutor = await prisma.tutor.findUnique({
+    where: { userId: user.userId }
+  });
+
+  if (!tutor) {
+    throw new AppError("Tutor not found", 404);
+  }
+
   return await prisma.tutorSubject.update({
     where: {
       id,
     },
-    data,
+    data: {
+      categoryId: payload.categoryId
+    },
     include: {
       category: true,
     },
   });
 };
 
-const deleteSubject = async (id: string) => {
+const deleteSubject = async (user: IRequestUser, id: string) => {
+  if(user.role !== "TUTOR") {throw new AppError("Only tutors can delete subject", 403)}
+
+  const tutor = await prisma.tutor.findUnique({
+    where: { userId: user.userId }
+  });
+
+  if (!tutor) {
+    throw new AppError("Tutor not found", 404);
+  }
+
+  const subject = await prisma.tutorSubject.findUnique({
+    where: {
+      id
+    }
+  })
+  if(!subject) {throw new AppError("Subject not found", 404)}
+
+  if(subject.tutorId !== tutor.id) {
+    throw new AppError("You can only delete your own subjects", 403)
+  }
+  
   return await prisma.tutorSubject.delete({
     where: {
       id,
